requirements.txt
fastapi
uvicorn
python-multipart
pydantic
torch
torchvision
Pillow
python-dotenv
jinja2
